services/vanna/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
groq==0.4.1
vanna==0.4.0
sqlalchemy==2.0.23
pandas==2.1.3
openai==1.3.5

